[NAYB-97] refactor : 장바구니, 장바구니 상품 Fixture 경로 변경
